+++ lib/gptk.rb
@@ -14,7 +14,7 @@ end
 # GPT Kit - A collection of useful tools for interacting with GPT agents and generating content
 module GPTK
   START_TIME = Time.now
-  VERSION = '0.17'.freeze
+  VERSION = '0.19'.freeze
 
   # Load configuration files
   Config.load_openai_setup